chore: rename the dashboard screen to statistics
